+++ projects/ids-enterprise-ng/src/lib/card/soho-card.module.ts
@@ -1,0 +1,22 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import {
+  SohoCardHeaderComponent,
+  SohoCardPaneComponent,
+  SohoCardComponent
+} from './soho-card.component';
+
+@NgModule({
+  imports: [CommonModule],
+  declarations: [
+    SohoCardHeaderComponent,
+    SohoCardPaneComponent,
+    SohoCardComponent,
+  ],
+  exports: [
+    SohoCardHeaderComponent,
+    SohoCardPaneComponent,
+    SohoCardComponent,
+  ]
+})
+export class SohoCardModule { }