+++ src/app/app.module.ts
@@ -43,6 +43,8 @@ import { ButtonDemoComponent } from './button/button.demo';
 import { CalendarDemoComponent } from './calendar/calendar.demo';
 import { CalendarLegendDemoComponent } from './calendar/calendar-legend.demo';
 import { CalendarUpdatedDemoComponent } from './calendar/calendar-updated.demo';
+import { CardsDemoComponent } from './cards/cards.demo';
+import { CardsExpandableDemoComponent } from './cards/cards-expandable.demo';
 import { ChartDemoComponent } from './chart/chart.demo';
 import { CheckBoxDemoComponent } from './checkbox/checkbox.demo';
 import { CirclepagerDemoComponent } from './circlepager/circlepager.demo';
@@ -286,6 +288,8 @@ import { DropdownMultiselectAttributesDemoComponent } from './dropdown/dropdown-
     CalendarDemoComponent,
     CalendarLegendDemoComponent,
     CalendarUpdatedDemoComponent,
+    CardsDemoComponent,
+    CardsExpandableDemoComponent,
     ChartDemoComponent,
     CheckBoxDemoComponent,
     CirclepagerDemoComponent,