+++ docs/CHANGELOG.md
@@ -7,11 +7,21 @@
 - `[General]` Temporary Placeholder
 - `[Multiselect]` Added an example showing automation id attributes on options in multiselect. ([#1005](https://github.com/infor-design/enterprise-ng/issues/1005))
 
+### 10.2.0 Features
+
+- `[Cards]` Added Angular Wrapper Component for the IDS Cards. ([#1036](https://github.com/infor-design/enterprise-ng/issues/1036)) `EA`
+
 ## v10.1.0
 
 ### 10.1.0 Fixes
 
 - `[General]` Upgraded the project to use Angular 12. See the [Angular docs](https://blog.angular.io/angular-v12-is-now-available-32ed51fbfd49) for more info on NG 12 also see the [UPGRADING.md](https://github.com/infor-design/enterprise-ng/blob/main/docs/UPGRADING.md) for steps we took to update (pretty simple this time). Note that support for IE 11 is removed in NG 12.
+
+## v9.5.3
+
+### 9.5.3 Fixes
+
+- `[General]` Remove tilde to bring in the right build
 
 ## v9.5.2
 
@@ -31,6 +41,12 @@
 
 - `[General]` Added EP version 4.51.0 final using NG 11.2.12.
 - `[StepProcess]` Removed unnecessary role=main from ts and template files. ([1033](https://github.com/infor-design/enterprise-ng/issues/1033)) `MHH`
+
+## v9.4.7
+
+### 9.4.7 Fixes
+
+- `[General]` Remove tilde to bring in the right build
 
 ## v9.4.5